moved config assignment code to config module
