+++ src/index.js
@@ -1,49 +1,23 @@
-#!/usr/bin/env node
+#!/usr/bin/env nodedev
 
 const args = process.argv;
 const fs = require("fs");
 const path = require('path');
 
 const postMethods = require("./posts");
-const config = require("./config");
+const configModule = require("./config");
+const config = configModule.config;
 const addHomePage = require("./homepage");
 
 const minArgv = require('minimist')(process.argv.slice(2));
 
-//chelc if it's just null or whitespace
-function isNullOrWhitespace( input ) {
-  return !input || !input.trim();
+
+var configFile=false;
+if(minArgv.config || minArgv.c) {
+  //flag for using config file 
+  configFile=configModule.setConfig(minArgv.config || minArgv.c);
+
 }
-
-//if config JSON file exist, save the content to config module
-if(minArgv.config || minArgv.c){
-  let argvConfig=minArgv.config || minArgv.c;
-  if (!fs.existsSync(argvConfig) || path.extname(argvConfig)!=".JSON") {
-    console.log("It's not an existing JSON file.");
-    return 0;
-  }
-  
-  //read the parse the config file
-  const configString = fs.readFileSync(argvConfig).toString();
-  try { 
-    var temptConfig = JSON.parse(configString);
-  }catch(err) {
-    console.log("Please make sure the JSON file is valid. Encountered an issue: "+err);
-    return 0;
-  }
-  
-  //checking if each option exist from the config file before assiging to confgi module
-  if(!isNullOrWhitespace(temptConfig.postsdir)) config.dev.postsdir = temptConfig.postsdir;
-  if(!isNullOrWhitespace(temptConfig.inputPath)) config.dev.inputPath = temptConfig.inputPath;
-  if(!isNullOrWhitespace(temptConfig.outdir)) config.dev.outdir = temptConfig.outdir;
-  if(!isNullOrWhitespace(temptConfig.lang)) config.dev.lang= temptConfig.lang;
-  if(!isNullOrWhitespace(temptConfig.stylesheet)) config.dev.stylesheet= temptConfig.stylesheet;
-
-  var configFile=true;//flag for using config file
-  
-}
-
-
 
 const posts = fs
           .readdirSync(config.dev.postsdir)
@@ -51,10 +25,9 @@ const posts = fs
           .map(post => postMethods.createPost(post))
 
 
-      
-//if config file is used, and the file path is valid, parse the input files
+  //if config file is used, and the file path is valid, parse the input files
 if(configFile && fs.existsSync(config.dev.postsdir)) {
-
+  
   //if it's a single .txt or .md file
   if (path.extname(config.dev.inputPath) == ".txt") {
     let p=path.join(config.dev.postsdir,config.dev.inputPath);
